builtin/run/binary: updates to satisfy run.Backend
This change updates the `Run` function to implement the `run.Backend` interface.
Additionally, this change switches from using `exec.Command` to `exec.CommandContext`
to enable the consuming package to cancel binary commands.

A tweak was made to use the `io.Writer`s specified in the `BackendConfig` versus
being set specifically to `os.Std*` for the `exec.Command`.