sdk/runner: adds a runner interface (#69)
* pkg/neighbor: removes old neighbor command

* sdk/run: defines a run factory interface

* builtin/run/binary: implements run interface

This change is a work-in-progress that adds an implementation of the runner
interface for running binary commands.

* sdk/run: adds params to Run()

This change adds a context.Context and dir string to the parameters
of the `Run` function.

* builtin/run/binary: updates to satisfy run.Backend

This change updates the `Run` function to implement the `run.Backend` interface.
Additionally, this change switches from using `exec.Command` to `exec.CommandContext`
to enable the consuming package to cancel binary commands.

A tweak was made to use the `io.Writer`s specified in the `BackendConfig` versus
being set specifically to `os.Std*` for the `exec.Command`.

* main: uses binary runner

This change is a first pass at using the `binary.Run`.

* docs: adds example for multi-line command

This changes add an example to the README with a multi-line command. It also acknowledges
neighbor's limitation to not currently handle piped commands.

* bulitin/run/binary: add Factory tests

This change adds a couple test cases for the `binary.Factory` function. It identifies
an existing bug where escaped strings (e.g., `grep \"FAIL\"`) are not respected and
get dropped. This leads to malformed commands being executed.

* builtin/run/binary: removes command scrubbing

This changes removes the `cleanWords` function that was called after parsing
commands and removed `"` and `'`. This function innappropriately removed these
characters.

* builtin/run/binary: fixes arg bug

This change fixes an off-by-one bug where if a command only had a single argument,
it was ignored. A test case has been added that covers this case.
Additionally, test cases were added to cover branches where errors are returned
in the `Factory` function.

* builtin/run/binary: adds tests for Run method

This change adds tests for and updates the `Run()` method.

* builtin/run/binary: adds GoDoc comments

This change adds comments to the `Factory()` function, the `Backend` struct and
`Run()` method.

* docs: updates README

This change updates the README multi-line command example.